implement model using keras layers (replacing tf.functions)

+++ dlsrl/utils.py
@@ -14,7 +14,7 @@ def shuffle_stack_pad(data, batch_size, shuffle=True):
     shape0_adj = shape0 - num_excluded
     shape1 = np.max([len(i) for i in data[1]])
 
-    mats = [np.zeros((shape0_adj, shape1)).astype(np.int32) for _ in range(3)]
+    mats = [np.zeros((shape0_adj, shape1)).astype(np.float32) for _ in range(3)]
     if shuffle:
         row_ids = np.random.choice(shape0, shape0_adj, replace=False)
     else:
